+++ changelog.md
@@ -6,6 +6,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 
 * chore: upgrade all dependencies to latest stable version
 * feat: adapt Jira style guide for Scrum Poker card layout
+* refactor: use Spring annotations and conform to atlassian-spring-scanner
 
 # [3.5.0] - 2018-09-16